requirements.txt
Django==1.11.5
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
pytz
sqlparse
mysqlclient==1.3.12             #required for sqlite databases in django
reportlab
django-smtp-ssl
django-widget-tweaks            #for add css class
pytest
pytest-django                   #pytest for django, prettier test reprots